Crudely implement essiv for aes cbc
Use sha256 as the key hashing algorithm to derive the salt from the given
aes key.

The iv can then be generated by encrypting the little endian representation
of the block id or block number with the hashed key using aes256 ecb.

Hardcode zero as the block id for now.