Begin implementing the basic ipld data model
Lay out the interface and the inheritance hierarchy needed to implement the
basic ipld data model.

Provide a trivial implementation for ipld nodes of kind null.

To reduce the amount of boilerplate code in implementations of the ipld
node interface, include a dummy base class whose members all throw an
exception on access. That way, derived classes implementing the ipld node
interface need not bother themselves with interface elements that are
unrelated to the kind of data they hold.