Create a basic implementation for PBLink
PBLink shall use an instance of PBLinkRepr to store the link data proper,
including the link name and the link target size. As it shall also serve as
an ipld data node, make it implement the INode interface.

In particular, implement the lookup by string for PBLink (as it acts as a
data node of kind map when cast to the INode interface).

PBLinkRepr shall become the PBLink interface to the world of protobuf later
on, to cleanly separate protobuf functionality from the data model.

Implement PBLink's child nodes as stubs that receive a reference to their
parent node's PBLinkRepr instance when they are created. That way, they can
directly modify the data in the PBLinkRepr instance relevant to them.