Link PBNode and PBLink without surrogates
Abandon the strict separation between the ipld data model and protobuf in
the inheritance hierarchy.

Move all relevant fields or members to PBNode and PBLink themselves, so
that protobuf-net can work on them without the need for surrogates.

As the classes representing various child nodes of PBNode or PBLink all
need a handle to their parents' data, create interfaces that expose the
necessary functionality to read or modify the parents' data.

PBNode shall deep copy all data provided via constructor parameters, if the
respective parameter is set.

Ensure that all list types used are generic .net lists and not f# lists, to
facilitate interoperability with other .net languages.