Prepare making ipld nodes immutable
Create an interface for future ipld node builders, and design the ipld node
interface in a way that appears consistent and idiomatic to users:

* getters shall, in general, act as pseudo-casts to the respective native
  .net data type. For example, the data contained in a node of integer kind
  shall be exposed as an uint32 for ToUInt32(), as an uint64 for ToUInt64()
  etc. while an exception shall be thrown if the native integral type can't
  hold the contained data so that an overflow occurs; and
* getters for recursive kinds shall return an object that can be upcast
  to an appropriate read-only .net collection interface (like IReadOnlyList
  in case of lists and IReadOnlyDictionary in case of maps).

Also, outline an extensible interface for ipld codecs so that they can be
easily customised by providing them with a custom node builder during
invocation. Create the interfaces IDecoder and IEncoder for future codecs
to implement to become usable with our evolving ipld library.