Outline the content id mapping storage mechanism
Begin by writing generic functions for lookup, insertion, update and
deletion operations. Extract common functionality into helper functions,
and make those helper functions generic as necessary.

Specify an interface for concrete storage mechanisms to implement.

Implementations shall be written in the form of classes that represent
individual nodes in the encrypted dag of the mapping store and implement
the previously mentioned interface. The interface exposes member functions
to construct new immutable mapping storage dags when provided with all
necessary parameters like keys and values to insert, update or delete as
well as the stack of nodes representing the path from the root node to the
node where the operation takes place.